+++ lib/src/provider.dart
@@ -1,23 +1,24 @@
 import 'package:flutter/material.dart';
+import 'package:sso/sso.dart';
 
 import 'authenticator.dart';
+import 'mixin/will_convert_user.dart';
 import 'mixin/will_notify.dart';
-import 'notifier.dart';
 import 'user.dart';
+import 'user_converter.dart';
 
-abstract class Provider {
-  Widget get header;
-
-  ThemeData get theme;
-
-  Notifier get notifier => Notifier();
-
+abstract class Provider with WillNotify, WillConvertUser {
   final List<Authenticator> _authenticators = List();
 
   /// Construct Provider
   ///
   /// Receives a list of authenticators as argument
-  Provider([List<Authenticator> authenticators = const []]) {
+  Provider(
+      [List<Authenticator> authenticators = const [],
+      Notifier notifier,
+      UserConverter userConverter]) {
+    this.notifier = notifier;
+    this.userConverter = userConverter;
     for (Authenticator authenticator in authenticators) {
       this.use(authenticator);
     }
@@ -27,6 +28,9 @@ abstract class Provider {
   Provider use(Authenticator authenticator) {
     if (authenticator is WillNotify) {
       (authenticator as WillNotify).notifier = notifier;
+    }
+    if (authenticator is WillConvertUser) {
+      (authenticator as WillConvertUser).userConverter = userConverter;
     }
     _authenticators.add(authenticator);
     return this;