+++ lib/sso.dart
@@ -1,8 +1,9 @@
 library sso;
 
-export 'src/action_button.dart';
 export 'src/authenticator.dart';
+export 'src/mixin/will_convert_user.dart';
 export 'src/mixin/will_notify.dart';
 export 'src/notifier.dart';
 export 'src/provider.dart';
 export 'src/user.dart';
+export 'src/user_converter.dart';